requirements.txt
streamlit
pandas
openpyxl
PyPDF2
python-docx
Pillow
python-magic
fpdf
PyQt5
xlwings